API: add droplevel() to flat indexes, for compatibility with MultiIndex
closes #21115
